fix openapi and rest endpoint conflicts
